readBits only can read for 1 byte but cant more than 1

+++ src/Read_File.h
@@ -9,7 +9,7 @@ typedef struct InStream InStream;
 struct InStream{
   FILE *file;
   char *filename;
-  uint8_t byteIndex;
+  uint32_t byteIndex;
   uint32_t bitIndex;
 };
